Do not import eth_tester until provider init
